02_coreml_conversion/requirements.txt
# CoreML 转换环境依赖
ultralytics>=8.0.0
coremltools>=7.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0